requirements.txt
matplotlib==3.8.0
numpy==1.26.0
odfpy==1.4.1
